still working player bug fix part 3
The most annoying part about this bug is that I have to run the program and wait for it to fail, it does not fail for at least half an hour. I upload these to github because I develop it on my PC but run it on my raspberry Pi 5. If it does not work on the Pi 5, I can't have it up 24/7.